# docker/swagger.yaml
swagger: "2.0"
info:
  description: "A server for a Cupcake store"
  version: "1.0.0"
  title: "Cupcake Store"
  termsOfService: "https://8flow.ai/"
  contact:
    email: "apiteam@cupcakestore.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/v2"
tags:
- name: "cupcake"
  description: "Everything about your Cupcakes"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "http"
paths:
  /cupcake:
    post:
      tags:
      - "cupcake"
      summary: "Add a new cupcake to the store"
      description: ""
      operationId: "addCupcake"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Cupcake object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Cupcake"
      responses:
        405:
          description: "Invalid input"
    get:
      tags:
      - "cupcake"
      summary: "List all cupcakes"
      description: "Returns a list of available cupcakes"
      operationId: "listCupcakes"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Cupcake"
  /cupcake/{cupcakeId}:
    get:
      tags:
      - "cupcake"
      summary: "Find cupcake by ID"
      description: "Returns a single cupcake"
      operationId: "getCupcakeById"
      produces:
      - "application/json"
      parameters:
      - name: "cupcakeId"
        in: "path"
        description: "ID of cupcake to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Cupcake"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cupcake not found"
    put:
      tags:
      - "cupcake"
      summary: "Update an existing cupcake"
      description: ""
      operationId: "updateCupcake"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "cupcakeId"
        in: "path"
        description: "ID of cupcake that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Updated cupcake object"
        required: true
        schema:
          $ref: "#/definitions/Cupcake"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cupcake not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "cupcake"
      summary: "Deletes a cupcake"
      description: ""
      operationId: "deleteCupcake"
      produces:
      - "application/json"
      parameters:
      - name: "cupcakeId"
        in: "path"
        description: "Cupcake id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cupcake not found"
definitions:
  Cupcake:
    type: "object"
    required:
    - "name"
    - "price"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      description:
        type: "string"
      price:
        type: "number"
      ingredients:
        type: "array"
        items:
          type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"